optimize: remove mustget

+++ server/cmd/api/biz/handler/car/car_service.go
@@ -52,9 +52,14 @@ func AdminDeleteCar(ctx context.Context, c *app.RequestContext) {
 		c.JSON(http.StatusBadRequest, resp)
 		return
 	}
-
+	aid, flag := c.Get(consts.AccountID)
+	if !flag {
+		resp.BaseResp = tools.BuildBaseResp(errno.AuthorizeFail)
+		c.JSON(http.StatusBadRequest, resp)
+		return
+	}
 	resp, err = config.GlobalCarClient.DeleteCar(ctx, &kcar.DeleteCarRequest{
-		AccountId: c.MustGet(consts.AccountID).(int64),
+		AccountId: aid.(int64),
 		Id:        req.ID,
 	})
 	if err != nil {
@@ -104,8 +109,13 @@ func AdminGetSomeCars(ctx context.Context, c *app.RequestContext) {
 		c.JSON(http.StatusBadRequest, resp)
 		return
 	}
-
-	resp, err = config.GlobalCarClient.GetSomeCars(ctx, &kcar.GetSomeCarsRequest{AccountId: c.MustGet(consts.AccountID).(int64)})
+	aid, flag := c.Get(consts.AccountID)
+	if !flag {
+		resp.BaseResp = tools.BuildBaseResp(errno.AuthorizeFail)
+		c.JSON(http.StatusBadRequest, resp)
+		return
+	}
+	resp, err = config.GlobalCarClient.GetSomeCars(ctx, &kcar.GetSomeCarsRequest{AccountId: aid.(int64)})
 	if err != nil {
 		hlog.Error("rpc car service err", err)
 		resp.BaseResp = tools.BuildBaseResp(errno.ServiceErr)
@@ -127,8 +137,13 @@ func AdminGetAllCars(ctx context.Context, c *app.RequestContext) {
 		c.JSON(http.StatusBadRequest, resp)
 		return
 	}
-
-	resp, err = config.GlobalCarClient.GetAllCars(ctx, &kcar.GetAllCarsRequest{AccountId: c.MustGet(consts.AccountID).(int64)})
+	aid, flag := c.Get(consts.AccountID)
+	if !flag {
+		resp.BaseResp = tools.BuildBaseResp(errno.AuthorizeFail)
+		c.JSON(http.StatusBadRequest, resp)
+		return
+	}
+	resp, err = config.GlobalCarClient.GetAllCars(ctx, &kcar.GetAllCarsRequest{AccountId: aid.(int64)})
 	if err != nil {
 		hlog.Error("rpc car service err", err)
 		resp.BaseResp = tools.BuildBaseResp(errno.ServiceErr)
@@ -150,8 +165,13 @@ func GetCars(ctx context.Context, c *app.RequestContext) {
 		c.JSON(http.StatusBadRequest, resp)
 		return
 	}
-
-	resp, err = config.GlobalCarClient.GetCars(ctx, &kcar.GetCarsRequest{AccountId: c.MustGet(consts.AccountID).(int64)})
+	aid, flag := c.Get(consts.AccountID)
+	if !flag {
+		resp.BaseResp = tools.BuildBaseResp(errno.AuthorizeFail)
+		c.JSON(http.StatusBadRequest, resp)
+		return
+	}
+	resp, err = config.GlobalCarClient.GetCars(ctx, &kcar.GetCarsRequest{AccountId: aid.(int64)})
 	if err != nil {
 		hlog.Error("rpc car service err", err)
 		resp.BaseResp = tools.BuildBaseResp(errno.ServiceErr)
@@ -173,9 +193,14 @@ func GetCar(ctx context.Context, c *app.RequestContext) {
 		c.JSON(http.StatusBadRequest, resp)
 		return
 	}
-
+	aid, flag := c.Get(consts.AccountID)
+	if !flag {
+		resp.BaseResp = tools.BuildBaseResp(errno.AuthorizeFail)
+		c.JSON(http.StatusBadRequest, resp)
+		return
+	}
 	resp, err = config.GlobalCarClient.GetCar(ctx, &kcar.GetCarRequest{
-		AccountId: c.MustGet(consts.AccountID).(int64),
+		AccountId: aid.(int64),
 		Id:        req.ID,
 	})
 	if err != nil {
